Произвел небольшой рефакторинг
