+++ app/src/main/java/ru/ktsstudio/wishlist/ui/main/MainNavigator.kt
@@ -1,5 +1,5 @@
 package ru.ktsstudio.wishlist.ui.main
 
 interface MainNavigator {
-    fun navigateToMain()
+    fun navigateToWishTabs()
 }
