+++ app/src/main/java/ru/ktsstudio/wishlist/ui/main/wishtabs/WishFragment.kt
@@ -1,5 +1,9 @@
 package ru.ktsstudio.wishlist.ui.main.wishtabs
 
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -10,7 +14,16 @@ import ru.ktsstudio.wishlist.ui.main.wishtabs.adapters.WishAdapter
 
 abstract class WishFragment : Fragment() {
 
-    fun setupRecyclerView() {
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_rv, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setupRecyclerView()
+    }
+
+    private fun setupRecyclerView() {
         val adapter = WishAdapter(getWishes())
         rv.adapter = adapter
         rv.layoutManager = LinearLayoutManager(activity)
